backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
psycopg2-binary>=2.9
drf-spectacular>=0.27
django-cors-headers>=4.3
